Refactor getSummaryDescriptionText.
