+++ app/component/itinerary/BoardingInformation.js
@@ -1,10 +1,6 @@
 import React from 'react';
 import cx from 'classnames';
 import PlatformNumber from '../PlatformNumber';
-import {
-  getTrackOrPierOrPlatformChangeText,
-  getTrackOrPierOrPlatformWithNumText,
-} from '../../util/localeUtils';
 import { isPlatformChanged } from '../../util/legUtils';
 import { legShape } from '../../util/shapes';
 
@@ -45,32 +41,4 @@ BoardingInformation.propTypes = {
   leg: legShape.isRequired,
 };
 
-/**
- * Returns a string with platform/track information for a transit leg, for screen reader use.
- * @param {Object} leg - The transit leg object.
- * @param {Object} intl - The intl object for formatting messages.
- * @returns {string} The boarding information text.
- */
-function getBoardingInformationText(leg, intl, showPlatformChangeLabel = true) {
-  if (!leg) {
-    return '';
-  }
-  const platformCode = leg?.from?.stop?.platformCode;
-  if (platformCode) {
-    const platformChangeLabelText =
-      showPlatformChangeLabel && isPlatformChanged(leg)
-        ? `${getTrackOrPierOrPlatformChangeText(intl, leg.mode)}:`
-        : '';
-    const platformLabel = getTrackOrPierOrPlatformWithNumText(
-      intl,
-      leg.mode,
-      platformCode,
-    );
-
-    return `${platformChangeLabelText} ${platformLabel}`;
-  }
-  return '';
-}
-
-export { getBoardingInformationText };
 export default BoardingInformation;
